set up font for using Japanese with vcd::mosaic

mosaic_jp*

Usage

mosaic_jp1()

References

- The Strucplot Framework: Visualizing Multi-way Contingency Tables with vcd: https://www.jstatsoft.org/index.php/jss/article/view/v017i03/v17i03.pdf

Examples

# installation

if(!require(devtools)) install.packages("devtools")
#> Loading required package: devtools
#> Loading required package: usethis
devtools::install_github("419kfj/futools")
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Downloading GitHub repo 419kfj/futools@HEAD
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpYsTXP1/remotes73a00327a51ed/419kfj-futools-6c15b39/DESCRIPTION’ ... OK
#> * preparing ‘futools’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘futools_0.1.3.3.tar.gz’
#> 
#> Installing package into ‘/tmp/RtmpB5SyjU/temp_libpath7377e17639b9c’
#> (as ‘lib’ is unspecified)

# data set
data(HairEyeColor_jp)

# plain mosaic_jp
mosaic_jp0(HairEyeColor_jp)
#> Error in mosaic_jp0(HairEyeColor_jp): could not find function "mosaic_jp0"

# without cell color
mosaic_jp2(HairEyeColor_jp)
#> Error in mosaic(.tbl, ..., margins = c(left = 5, top = 4), rot_labels = c(left = 0,     top = 45, right = 0), keep_aspect_ratio = FALSE, main = mtitle,     labeling = labeling_values): could not find function "mosaic"

# set cell color by color matrix
cset <- t(matrix(rep(brewer.pal(4,"Blues"),2),ncol=2))
#> Error in brewer.pal(4, "Blues"): could not find function "brewer.pal"
mosaic_jp2(HairEyeColor_jp,gp=gpar(fill=cset,col=0))
#> Error in mosaic(.tbl, ..., margins = c(left = 5, top = 4), rot_labels = c(left = 0,     top = 45, right = 0), keep_aspect_ratio = FALSE, main = mtitle,     labeling = labeling_values): could not find function "mosaic"

# Pearson residulas
mosaic_jp2(HairEyeColor_jp,shade=TRUE)
#> Error in mosaic(.tbl, ..., margins = c(left = 5, top = 4), rot_labels = c(left = 0,     top = 45, right = 0), keep_aspect_ratio = FALSE, main = mtitle,     labeling = labeling_values): could not find function "mosaic"